requirements-dev.txt
numpy 
pandas 
python-dotenv 
snowflake-sqlalchemy
sqlalchemy 
plotly 
streamlit 
openai==0.28.0 
boto3 
snowflake-snowpark-python 
langchain 
langchain-experimental